Move useSubscriptionDetails to shared hooks dir
