+++ src/components/users/subscription-details.tsx
@@ -1,49 +1,13 @@
 import React from 'react'
-import axios from 'axios'
 import Link from 'next/link'
 import {track} from '../../utils/analytics'
 import {useViewer} from 'context/viewer-context'
 import get from 'lodash/get'
+import useSubscriptionDetails from 'hooks/use-subscription-data'
 
 type SubscriptionDetailsProps = {
   stripeCustomerId: string
   slug: string
-}
-
-const useSubscriptionDetails = ({
-  stripeCustomerId,
-  slug,
-}: SubscriptionDetailsProps) => {
-  const [subscriptionData, setSubscriptionData] = React.useState<any>()
-  const recur = (price: any) => {
-    const {
-      recurring: {interval, interval_count},
-    } = price
-
-    if (interval === 'month' && interval_count === 3) return 'quarter'
-    if (interval === 'month' && interval_count === 6) return '6-months'
-    if (interval === 'month' && interval_count === 1) return 'month'
-    if (interval === 'year' && interval_count === 1) return 'year'
-  }
-
-  React.useEffect(() => {
-    if (stripeCustomerId) {
-      axios
-        .get(`/api/stripe/billing/session`, {
-          params: {
-            customer_id: stripeCustomerId,
-            account_slug: slug,
-          },
-        })
-        .then(({data}) => {
-          if (data) {
-            setSubscriptionData(data)
-          }
-        })
-    }
-  }, [stripeCustomerId, slug])
-
-  return [subscriptionData, recur]
 }
 
 const SubscriptionDetails: React.FunctionComponent<SubscriptionDetailsProps> = ({
